uc/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiohttp==3.9.1
redis==5.0.1
python-json-logger==2.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
cryptography==41.0.7